// src/utils/interfaces/payouyTransaction.interface.ts
export type IPayoutStatistics = {
    earnings: number;
    earningsFees: number;
    success: number;
    failed: number;
    successRate: number;
    returned: number;
    initiate: number;
    total: number;
    balance: number;
    payoutBalance: number;
    payinBalance: number;
    settlementBalance: number;
};